refactor: reorganize file system models and update imports for consistency
